New synthesizer code has been renamed 'synthesizer_grpc'; old synthesizer code has been restored
